Add database tables and APIs for tracking user's skills and desired skills
These skills can be added, removed or updated while creating the user,
during onboarding, or later by editing the user profile.

If a specified skill is not already present in the database, it will be
added as a new "custom" skill. Skills can have expertise levels, but
desired skills don't have any levels.