Update skills API to return 20 skills by default
This will allow the frontend to remove skills that have already
been selected, while still showing at least 10 skills.

Also added a "limit" parameter to the query string to let the caller
control how many skills should be returned. If the limit is set to 0,
the default number of skills are returned. No more than 50 skills will
be returned regardless of the parameter.